[2.1.0] Allow automatic seeding after automatic migrations
Fixes #158

If config('tenancy.seeder_class') is null or a blank string, the seeder-class parameter is not passed, and therefore Laravel's default option of calling 'DatabaseSeeder' will occur.

As with Laravel's normal seeding process, the class is treated as the "root" seeder class, which can call other seeder classes in turn.