Update UserImpersonation.php
This PR adds a new session value `tenancy_impersonated_by` in the tenant scope when you are impersonating a tenant.

This is similar to how Laravel Nova adds the `nova_impersonated_by` session value and could be used to identify when you are impersonating, who is impersonating. This may be useful for additional application logic and UI niceties such as an impersonation banner.