Make morph maps work with resource syncing
Before, morph maps didn't work because in TriggerSyncingTrait,
we were trying to get the pivot's `$this->morphClass`. Now,
we're first checking if the morph class is an alias, and if it is, resolve
its full class name, if it is not, `$this->morphClass` is already
the class name, so we can use it.